# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: forge-dashboard
  summary: A single-page view of open PRs, issues and CI status across GitHub and Forgejo.
  description: |
    Aggregates every repository the configured GitHub and Forgejo accounts
    have write access to into one read-only view: open pull requests (with
    CI status), open issues, and per-forge reachability. The backend holds
    both forges' credentials; the frontend only ever calls the endpoints
    below.
  version: 1.0.0
  contact:
    name: forge-dashboard
    url: https://github.com/alrayyes/forge-dashboard
  license:
    name: AGPL-3.0
    identifier: AGPL-3.0-only
servers:
  - url: "{scheme}://{host}"
    description: |
      No canonical host — this runs on a homelab reachable only over
      Tailscale, so the variables below are the shape of the answer rather
      than the answer.
    variables:
      scheme:
        default: https
        enum: [https, http]
      host:
        default: localhost:8080
        description: The default listen address. Set ADDR to change it.
security: []
...
paths:
...
  /api/pull-requests/merge:
    post:
      operationId: mergePullRequest
      summary: Merge one pull request, on the signed-in user's behalf
      description: |
        Merges the named pull request using its repo's own configured
        default merge method — GitHub is asked to pick its own repo
        default; Forgejo's API has no such default built in, so this
        looks up the repo's configured default merge style first and
        passes that explicitly. Neither takes a method override here;
        picking one is out of scope for this endpoint.

        When the forge refuses the merge, the server re-reads the pull
        request's real current state and answers an `ActionError`: a `code`
        from a fixed set plus a short `message` safe to show a person, so
        every client gets the same reason without parsing forge text. A
        pull request that is already merged or closed answers
        `already_merged` / `already_closed`. If the re-read itself fails,
        the original forge error is returned with `code: unknown`.
      tags: [pull-requests]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullRequestActionRequest"
      responses:
        "204":
          description: Merged.
        "400":
          description: |
            The request body wasn't valid JSON, `fullName` wasn't
            "owner/repo", no credentials are saved for that forge, or
            that forge's client doesn't support merging pull requests at
            all.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No valid session cookie was presented.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: The forge rejected the request as unauthorized — the saved token's scope doesn't cover merging, or it isn't allowed to merge on that repo. `code` is `permission`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "404":
          description: The forge reported the repo or pull request doesn't exist (or isn't visible to this token). `code` is `unknown`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "409":
          description: |
            The forge refused the merge and the pull request's current
            state says why. `code` is one of `already_merged`,
            `already_closed`, `not_mergeable`, `conflict`, `behind`,
            `checks_pending`, `checks_failing` or `blocked_by_protection`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "429":
          description: The forge rate-limited the request. `code` is `rate_limited`; `resetsAt` carries when the budget comes back, when the forge said.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "502":
          description: The forge was unreachable, or answered with something this app couldn't classify, or the follow-up re-read failed. `code` is `unknown`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
...
  /api/pull-requests/close:
    post:
      operationId: closePullRequest
      summary: Close one pull request without merging it, on the signed-in user's behalf
      description: |
        Closes the named pull request — for one that turns out not to
        need merging at all (a duplicate, or one whose content already
        landed another way), not a substitute for Merge.

        When the forge refuses, the server re-reads the pull request and
        answers an `ActionError` (see Merge): `already_merged` or
        `already_closed` when the row was stale, otherwise a `code` and a
        plain-words `message` safe to show a person.
        Closing a merged pull request is `already_merged`.
      tags: [pull-requests]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullRequestActionRequest"
      responses:
        "204":
          description: Closed.
        "400":
          description: |
            The request body wasn't valid JSON, `fullName` wasn't
            "owner/repo", no credentials are saved for that forge, or
            that forge's client doesn't support closing pull requests at
            all.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No valid session cookie was presented.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: The forge rejected the request as unauthorized — the saved token's scope doesn't cover closing, or it isn't allowed to close on that repo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "404":
          description: The forge reported the repo or pull request doesn't exist (or isn't visible to this token). `code` is `unknown`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "429":
          description: The forge rate-limited the request. `code` is `rate_limited`; `resetsAt` carries when the budget comes back, when the forge said.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
        "502":
          description: The forge was unreachable, or answered with something this app couldn't classify more specifically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionError"
...
components:
  securitySchemes:
    # Applied via a `security:` block on every operation that accepts it
    # server-side (internal/api/server.go's auth.RequireAuth wiring) — see
    # the auth tag's own description for why an endpoint's 401 description
    # says "no valid session cookie" without separately calling out this
    # header: both credentials produce the same 401. The two webhook
    # delivery endpoints are the deliberate exception; see their own
    # `security: []` comments.
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal API token from POST /api/tokens, as an alternative to the session cookie.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ActionError:
      type: object
      required: [error, code, message]
      description: >-
        The structured result of a refused pull request action (Merge,
        Close, Update branch, Enable auto-merge, Dependabot and Renovate
        rebase). `error` is the same
        string every Error carries (the forge's own text, for logs);
        `code` and `message` are what a client should act on and show.
      properties:
        error:
          type: string
          description: The underlying error text, unchanged.
        code:
          type: string
          enum:
            - already_merged
            - already_closed
            - not_mergeable
            - conflict
            - behind
            - checks_pending
            - checks_failing
            - blocked_by_protection
            - already_up_to_date
            - auto_merge_not_allowed
            - ready_to_merge
            - permission
            - rate_limited
            - unknown
          description: |
            Why the action was refused, from a re-read of the pull
            request's real state. `already_merged` and `already_closed`
            mean the dashboard's row was stale: the pull request has
            nothing left to merge.

            Three codes belong to one action each: `already_up_to_date`
            (Update branch: nothing to bring in), `auto_merge_not_allowed`
            (Enable auto-merge: the repo or pull request doesn't allow it)
            and `ready_to_merge` (Enable auto-merge: already clean, use
            Merge).
        message:
          type: string
          description: A short reason in plain words, safe to show a person.
        resetsAt:
          type: string
          format: date-time
          description: Only with `rate_limited`, when the forge said so. When the budget comes back.
...
    PullRequestActionRequest:
      type: object
      required: [forge, fullName, number]
      description: Which pull request to act on.
      properties:
        forge:
          $ref: "#/components/schemas/Forge"
        fullName:
          type: string
          description: '"owner/repo", matching a Repo.fullName from GET /api/dashboard.'
        number:
          type: integer
          description: Matching a PullRequest.number from GET /api/dashboard.
...
    Forge:
      type: string
      enum: [github, forgejo]
...