# internal/api/pull_request_close.go
package api

import (
	"encoding/json"
	"log/slog"
	"net/http"

	"github.com/alrayyes/forge-dashboard/internal/auth"
	"github.com/alrayyes/forge-dashboard/internal/dashboard"
)

// handlePullRequestClose closes the named pull request without merging
// it — dashboard.PullRequestCloser's own ClosePullRequest decides how,
// per forge. Follows the same shape handlePullRequestMerge already
// established.
func handlePullRequestClose(deps Deps) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		u, ok := auth.UserFromContext(r.Context())
		if !ok {
			writeJSON(w, http.StatusInternalServerError, errorBody("no authenticated user in context"))

			return
		}

		var req pullRequestActionRequest
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			writeJSON(w, http.StatusBadRequest, errorBody("invalid request body"))

			return
		}

		owner, name, ok := splitFullName(req.FullName)
		if !ok {
			writeJSON(w, http.StatusBadRequest, errorBody(`fullName must be "owner/repo"`))

			return
		}

		creds, err := deps.SettingsStore.Get(r.Context(), u.ID)
		if err != nil {
			writeJSON(w, http.StatusInternalServerError, errorBody("could not load settings"))

			return
		}

		var closer dashboard.PullRequestCloser
		for _, src := range deps.BuildSources(u.ID, creds) {
			if string(src.Forge()) != req.Forge {
				continue
			}
			c, supported := src.(dashboard.PullRequestCloser)
			if !supported {
				writeJSON(w, http.StatusBadRequest, errorBody(req.Forge+" doesn't support closing pull requests"))

				return
			}
			closer = c

			break
		}
		if closer == nil {
			writeJSON(w, http.StatusBadRequest, errorBody("no "+req.Forge+" credentials saved"))

			return
		}

		if err := closer.ClosePullRequest(r.Context(), owner, name, req.Number); err != nil {
			slog.Warn("pull request close failed", "forge", req.Forge, "repo", req.FullName, "number", req.Number, "error", err)
			writeActionRefusal(r.Context(), w, closer, dashboard.PullRequestActionClose, owner, name, req.Number, err)

			return
		}

		w.WriteHeader(http.StatusNoContent)
	}
}